Move core inventory code to inventory.js
