+++ inventory.js
@@ -5,6 +5,9 @@ import {RigAux} from './rig-aux.js';
 import runtime from './runtime.js';
 import {addDefaultLights} from './util.js';
 
+const localVector = new THREE.Vector3();
+const localRaycaster = new THREE.Raycaster();
+
 const inventorySpecs = [
   {
     "start_url": "https://ipfs.exokit.org/QmWgij93j5oU9SHGtGcKATJVsQ5ZEarqjg3bZDZMxXLdjF/model.vrm",
@@ -264,9 +267,12 @@ const planeMesh = new THREE.Mesh(new THREE.CylinderBufferGeometry(1, 1, 0.1).app
 planeMesh.position.set(0, -0.2, -2.5);
 inventoryAvatarScene.add(planeMesh);
 
-// XXX
+
+let avatarMeshLoaded = false;
 let avatarMesh = null;
-(async () => {
+let targetPoint = new THREE.Vector3(0, 0, 0);
+const initialAvatarPosition = new THREE.Vector3(0, 1.3, -2.5);
+const _loadAvatarMesh = async () => {
   const url = `https://webaverse.github.io/assets/male.vrm`;
   const name = 'male.vrm';
   avatarMesh = await runtime.loadFile({
@@ -289,10 +295,107 @@ let avatarMesh = null;
     
     inventoryAvatarScene.add(rig.model);
   }
-})();
+};
+const _ensureAvatarMesh = () => {
+  if (!avatarMeshLoaded) {
+    avatarMeshLoaded = true;
+    _loadAvatarMesh();
+  }
+};
+
+const inventoryMenuEl = document.getElementById('inventory-menu');
+const inventoryAvatarEl = document.getElementById('inventory-avatar');
+
+const _isOpen = () => inventoryMenuEl.classList.contains('open');
+const toggle = () => {
+  inventoryMenuEl.classList.toggle('open');
+  document.exitPointerLock();
+};
+
+window.addEventListener('mousemove', e => {
+  if (_isOpen()) {
+    const boundingBox = inventoryAvatarRenderer.domElement.getBoundingClientRect();
+
+    const mouse = new THREE.Vector2();
+    mouse.x = ((e.clientX - (boundingBox.left)) / boundingBox.width) * 2 - 1;
+    mouse.y = - ((e.clientY - boundingBox.top) / boundingBox.height) * 2 + 1;
+    if (isFinite(mouse.x) && isFinite(mouse.y)) {
+      // console.log('box', mouse.toArray());
+      localRaycaster.setFromCamera(mouse, inventoryAvatarCamera);
+
+      if (avatarMesh) {
+        targetPoint = new THREE.Vector3(0, inventoryAvatarCamera.position.y, -1.5).applyQuaternion(
+          new THREE.Quaternion().setFromUnitVectors(
+            new THREE.Vector3(0, 0, -1),
+            localRaycaster.ray.direction
+          )
+        );
+        // cubeMesh.position.copy(targetPoint);
+
+        avatarMesh.rig.inputs.hmd.position.copy(initialAvatarPosition)
+          .add(new THREE.Vector3(mouse.x * 0.03, mouse.y * 0.1, 0));
+        avatarMesh.rig.inputs.hmd.quaternion.setFromRotationMatrix(new THREE.Matrix4().lookAt(
+          avatarMesh.rig.inputs.hmd.position,
+          targetPoint,
+          new THREE.Vector3(0, 1, 0)
+        ));
+      }
+    }
+  }
+});
+
+const update = () => {
+  if (_isOpen()) {
+    _ensureAvatarMesh();
+
+    inventoryAvatarCamera.position.set(0, 0.8, 0);
+
+    avatarMesh = inventoryAvatarScene.getObjectByName('avatarMesh');
+    if (avatarMesh) {
+      // avatarMesh.rig.inputs.hmd.position.copy(initialAvatarPosition);
+      /* avatarMesh.rig.inputs.hmd.quaternion.setFromUnitVectors(
+        new THREE.Vector3(0, 0, -1),
+        new THREE.Vector3(0, 0, 1)
+      ); */
+      {
+        avatarMesh.rig.inputs.leftGamepad.position.copy(avatarMesh.rig.inputs.hmd.position).add(localVector.set(-0.25, -0.3, 0.1));
+        const direction = targetPoint.clone().sub(avatarMesh.rig.inputs.leftGamepad.position).normalize();
+        avatarMesh.rig.inputs.leftGamepad.position.add(new THREE.Vector3(direction.x*0.5, direction.y*0.5, direction.z*0.25));
+        avatarMesh.rig.inputs.leftGamepad.quaternion.setFromRotationMatrix(new THREE.Matrix4().lookAt(
+          new THREE.Vector3(0, 0, 0),
+          direction,
+          new THREE.Vector3(0, 1, 0)
+        ))//.multiply(new THREE.Quaternion().setFromAxisAngle(new THREE.Vector3(0, 1, 0), Math.PI/2));
+        // avatarMesh.rig.inputs.leftGamepad.grip = 1;
+        // avatarMesh.rig.inputs.leftGamepad.pointer = 1;
+        // avatarMesh.rig.inputs.leftGamepad.enabled = true;
+      }
+      {
+        avatarMesh.rig.inputs.rightGamepad.position.copy(avatarMesh.rig.inputs.hmd.position).add(localVector.set(0.25, -0.8, 0.05));
+        const direction = targetPoint.clone().sub(avatarMesh.rig.inputs.rightGamepad.position).normalize();
+        // avatarMesh.rig.inputs.rightGamepad.position.add(direction.clone().multiplyScalar(0.3));
+        avatarMesh.rig.inputs.rightGamepad.quaternion.setFromRotationMatrix(new THREE.Matrix4().lookAt(
+          new THREE.Vector3(0, 0, 0),
+          new THREE.Vector3(1, 0, 0),
+          new THREE.Vector3(0, 0, 1)
+        ));
+        // avatarMesh.rig.inputs.rightGamepad.grip = 1;
+        // avatarMesh.rig.inputs.rightGamepad.pointer = 1;
+        // avatarMesh.rig.inputs.rightGamepad.enabled = true;
+      }
+      inventoryAvatarRenderer.render(inventoryAvatarScene, inventoryAvatarCamera);
+
+      const timeDiff = 0.01;
+      avatarMesh.rig.update(timeDiff);
+      avatarMesh.rig.aux.update(timeDiff);
+    }
+  }
+};
 
 export {
   inventoryAvatarScene,
   inventoryAvatarCamera,
   inventoryAvatarRenderer,
+  toggle,
+  update,
 };
