+++ weapons-manager.js
@@ -15,7 +15,7 @@ import {buildMaterial} from './shaders.js';
 import {makeTextMesh} from './vr-ui.js';
 import {teleportMeshes} from './teleport.js';
 import {appManager, renderer, scene, orthographicScene, camera, dolly} from './app-object.js';
-import {inventoryAvatarScene, inventoryAvatarCamera, inventoryAvatarRenderer} from './inventory.js';
+import {inventoryAvatarScene, inventoryAvatarCamera, inventoryAvatarRenderer, update as inventoryUpdate} from './inventory.js';
 import buildTool from './build-tool.js';
 import * as notifications from './notifications.js';
 import * as popovers from './popovers.js';
@@ -499,9 +499,6 @@ const _ungrab = () => {
   _updateMenu();
 };
 
-let avatarMesh = null;
-let targetPoint = new THREE.Vector3(0, 0, 0);
-const initialAvatarPosition = new THREE.Vector3(0, 1.3, -2.5);
 const crosshairEl = document.querySelector('.crosshair');
 const _updateWeapons = () => {  
   const transforms = rigManager.getRigTransforms();
@@ -827,49 +824,7 @@ const _updateWeapons = () => {
 
   popovers.update();
 
-  if (weaponsManager.inventoryOpen) {
-    inventoryAvatarCamera.position.set(0, 0.8, 0);
-
-    avatarMesh = inventoryAvatarScene.getObjectByName('avatarMesh');
-    if (avatarMesh) {
-      // avatarMesh.rig.inputs.hmd.position.copy(initialAvatarPosition);
-      /* avatarMesh.rig.inputs.hmd.quaternion.setFromUnitVectors(
-        new THREE.Vector3(0, 0, -1),
-        new THREE.Vector3(0, 0, 1)
-      ); */
-      {
-        avatarMesh.rig.inputs.leftGamepad.position.copy(avatarMesh.rig.inputs.hmd.position).add(localVector.set(-0.25, -0.3, 0.1));
-        const direction = targetPoint.clone().sub(avatarMesh.rig.inputs.leftGamepad.position).normalize();
-        avatarMesh.rig.inputs.leftGamepad.position.add(new THREE.Vector3(direction.x*0.5, direction.y*0.5, direction.z*0.25));
-        avatarMesh.rig.inputs.leftGamepad.quaternion.setFromRotationMatrix(new THREE.Matrix4().lookAt(
-          new THREE.Vector3(0, 0, 0),
-          direction,
-          new THREE.Vector3(0, 1, 0)
-        ))//.multiply(new THREE.Quaternion().setFromAxisAngle(new THREE.Vector3(0, 1, 0), Math.PI/2));
-        // avatarMesh.rig.inputs.leftGamepad.grip = 1;
-        // avatarMesh.rig.inputs.leftGamepad.pointer = 1;
-        // avatarMesh.rig.inputs.leftGamepad.enabled = true;
-      }
-      {
-        avatarMesh.rig.inputs.rightGamepad.position.copy(avatarMesh.rig.inputs.hmd.position).add(localVector.set(0.25, -0.8, 0.05));
-        const direction = targetPoint.clone().sub(avatarMesh.rig.inputs.rightGamepad.position).normalize();
-        // avatarMesh.rig.inputs.rightGamepad.position.add(direction.clone().multiplyScalar(0.3));
-        avatarMesh.rig.inputs.rightGamepad.quaternion.setFromRotationMatrix(new THREE.Matrix4().lookAt(
-          new THREE.Vector3(0, 0, 0),
-          new THREE.Vector3(1, 0, 0),
-          new THREE.Vector3(0, 0, 1)
-        ));
-        // avatarMesh.rig.inputs.rightGamepad.grip = 1;
-        // avatarMesh.rig.inputs.rightGamepad.pointer = 1;
-        // avatarMesh.rig.inputs.rightGamepad.enabled = true;
-      }
-      inventoryAvatarRenderer.render(inventoryAvatarScene, inventoryAvatarCamera);
-
-      const timeDiff = 0.01;
-      avatarMesh.rig.update(timeDiff);
-      avatarMesh.rig.aux.update(timeDiff);
-    }
-  }
+  inventoryUpdate();
 };
 
 /* const cubeMesh = new THREE.Mesh(new THREE.BoxBufferGeometry(0.01, 0.01, 0.01), new THREE.MeshBasicMaterial({
@@ -877,38 +832,6 @@ const _updateWeapons = () => {
   side: THREE.DoubleSide,
 }));
 inventoryAvatarScene.add(cubeMesh); */
-
-window.addEventListener('mousemove', e => {
-  if (weaponsManager.inventoryOpen) {
-    const boundingBox = inventoryAvatarRenderer.domElement.getBoundingClientRect();
-
-    const mouse = new THREE.Vector2();
-    mouse.x = ((e.clientX - (boundingBox.left)) / boundingBox.width) * 2 - 1;
-    mouse.y = - ((e.clientY - boundingBox.top) / boundingBox.height) * 2 + 1;
-    if (isFinite(mouse.x) && isFinite(mouse.y)) {
-      // console.log('box', mouse.toArray());
-      localRaycaster.setFromCamera(mouse, inventoryAvatarCamera);
-
-      if (avatarMesh) {
-        targetPoint = new THREE.Vector3(0, inventoryAvatarCamera.position.y, -1.5).applyQuaternion(
-          new THREE.Quaternion().setFromUnitVectors(
-            new THREE.Vector3(0, 0, -1),
-            localRaycaster.ray.direction
-          )
-        );
-        // cubeMesh.position.copy(targetPoint);
-
-        avatarMesh.rig.inputs.hmd.position.copy(initialAvatarPosition)
-          .add(new THREE.Vector3(mouse.x * 0.03, mouse.y * 0.1, 0));
-        avatarMesh.rig.inputs.hmd.quaternion.setFromRotationMatrix(new THREE.Matrix4().lookAt(
-          avatarMesh.rig.inputs.hmd.position,
-          targetPoint,
-          new THREE.Vector3(0, 1, 0)
-        ));
-      }
-    }
-  }
-});
 
 /* renderer.domElement.addEventListener('wheel', e => {
   if (document.pointerLockElement) {
@@ -1890,13 +1813,6 @@ const weaponsManager = {
     const auxPose = rigManager.localRig.aux.getPose();
     auxPose.sittables.length = 0;
     rigManager.localRig.aux.setPose(auxPose);
-  },
-  inventoryOpen() {
-    return inventoryMenuEl.classList.contains('open');
-  },
-  toggleInventory() {
-    inventoryMenuEl.classList.toggle('open');
-    document.exitPointerLock();
   },
   update() {
     _updateWeapons();
