+++ world.js
@@ -573,25 +573,3 @@ messages.addEventListener('messageadd', e => {
     }));
   }
 });
-
-(async () => {
-  const url = `https://webaverse.github.io/assets/male.vrm`;
-  const name = 'male.vrm';
-  const mesh = await runtime.loadFile({
-    url,
-    name,
-  });
-  if (mesh) {
-    mesh.name = 'avatarMesh';
-    const localRig = new Avatar(mesh.raw, {
-      fingers: true,
-      hair: true,
-      visemes: true,
-      debug: false //!o,
-    });
-    localRig.aux = new RigAux(localRig);
-    localRig.model.rig = localRig;
-    
-    inventoryAvatarScene.add(localRig.model);
-  }
-})();
